pyproject.toml
[tool.poetry]
name = "yazio"
version = "0.1.0"
description = ""
authors = [""]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
appium-python-client = "^4.4.0"
requests = "^2.32.3"
selene = "2.0.0rc9"
allure-pytest = "^2.13.5"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
